Add container and clean code a bit
