+++ app/actions/index.js
@@ -1,23 +1,3 @@
-export const fetchMovieData = (url) => {
-  return (dispatch) => {
-    dispatch(movieFetchLoading(true));
-
-    fetch(url)
-      .then(response => {
-        if(!response.ok) {
-          throw Error(response.statusText);
-        }
-
-        dispatch(movieFetchLoading(false));
-
-        return response;
-      })
-      .then(response => response.json())
-      .then(data => dispatch(movieFetchSuccess(data.results)))
-      .catch( () => dispatch(movieFetchError(true)))
-  }
-}
-
 export const movieFetchError = (bool) => {
     return {
         type: 'MOVIE_FETCH_ERROR',
@@ -37,4 +17,24 @@ export const movieFetchSuccess = (movies) => {
         type: 'MOVIE_FETCH_SUCCESS',
         data: movies
     };
+}
+
+export const fetchMovieData = (url) => {
+  return (dispatch) => {
+    dispatch(movieFetchLoading(true));
+
+    fetch(url)
+      .then(response => {
+        if(!response.ok) {
+          throw Error(response.statusText);
+        }
+
+        dispatch(movieFetchLoading(false));
+
+        return response;
+      })
+      .then(response => response.json())
+      .then(data => dispatch(movieFetchSuccess(data.results)))
+      .catch( () => dispatch(movieFetchError(true)))
+  }
 }
