+++ app/containers/AppContainer.js
@@ -1,19 +1,19 @@
-// import { connect } from 'react-redux';
-// import { fetchMovieData } from '../actions/index';
-// import App from '../components/App/App';
-//
-// const mapStateToProps(state) => {
-//   return {
-//     data: state.movies,
-//     hasErrored: state.fetchError,
-//     isLoading: state.fetchLoading
-//   }
-// }
-//
-// const mapDispatchToProps = (dispatch) => {
-//   return {
-//     fetchData: (url) => dispatch(fetchMovieData(url))
-//   };
-// }
-//
-// export default connect(mapStateToProps, mapDispatchToProps)(App)
+import { connect } from 'react-redux';
+import { fetchMovieData } from '../actions/index';
+import App from '../components/App/App';
+
+const mapStateToProps = (state) => {
+  return {
+    data: state.movies,
+    hasErrored: state.fetchError,
+    isLoading: state.fetchLoading
+  }
+}
+
+const mapDispatchToProps = (dispatch) => {
+  return {
+    fetchData: (url) => dispatch(fetchMovieData(url))
+  };
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(App)
